refactor: extract useCollapsibleHeight hook
Shared hook for animating height expand/collapse transitions,
replacing duplicated rAF logic in CollapsibleText and AccordionCard.
Also fixes missing rAF cleanup in CollapsibleText.
